+++ priv/templates/api/confirm_test.exs
@@ -14,14 +14,14 @@ defmodule <%= base %>.ConfirmTest do
   # In this case, you need to add the confirmation_token (key)
   # to a user with the email "gladys@mail.com" before running these tests
   test "confirmation succeeds for correct key" do
-    conn = conn() |> get("/confirm?" <> @valid_link)
+    conn = build_conn() |> get("/confirm?" <> @valid_link)
     assert conn.private.openmaize_info =~ "successfully confirmed"
     assert response(conn, 200)
     assert redirected_to(conn) == "/login"
   end
 
   test "confirmation fails for incorrect key" do
-    conn = conn() |> get("/confirm?" <> @invalid_link)
+    conn = build_conn() |> get("/confirm?" <> @invalid_link)
     assert conn.private.openmaize_error =~ "failed"
     assert response(conn, 200)
   end
@@ -30,19 +30,19 @@ defmodule <%= base %>.ConfirmTest do
   # You need to add the reset_token (key) to a user
   # with the email "gladys@mail.com" before running these tests
   test "reset password succeeds for correct key" do
-    conn = conn() |> post("/reset", user: @valid_attrs)
+    conn = build_conn() |> post("/reset", user: @valid_attrs)
     assert conn.private.openmaize_info =~ "successfully confirmed"
     assert response(conn, 200)
   end
 
   test "reset password fails for incorrect key" do
-    conn = conn() |> post("/reset", user: @invalid_attrs)
+    conn = build_conn() |> post("/reset", user: @invalid_attrs)
     assert conn.private.openmaize_error =~ "failed"
     assert response(conn, 200)
   end
 
   test "reset password fails for invalid password" do
-    conn = conn() |> post("/reset", user: @invalid_pass)
+    conn = build_conn() |> post("/reset", user: @invalid_pass)
     assert conn.private.openmaize_error =~ "password should be"
     assert response(conn, 200)
   end