+++ test/openmaize/confirm_email_test.exs
@@ -3,7 +3,7 @@ defmodule Openmaize.ConfirmEmailTest do
   use Plug.Test
 
   import Ecto.Changeset
-  alias Openmaize.{ConfirmEmail, TestRepo, User}
+  alias Openmaize.{ConfirmEmail, TestRepo, TestUser}
 
   @valid_link "email=fred%2B1%40mail.com&key=lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw"
   @name_link "username=fred&key=lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw"
@@ -11,9 +11,9 @@ defmodule Openmaize.ConfirmEmailTest do
   @incomplete_link "email=wrong%40mail.com"
 
   setup do
-    {:ok, _user} = TestRepo.get_by(User, email: "fred+1@mail.com")
+    {:ok, _user} = TestRepo.get_by(TestUser, email: "fred+1@mail.com")
     |> change(%{confirmed_at: nil})
-    |> Openmaize.Config.repo.update
+    |> Openmaize.TestRepo.update
     :ok
   end
 
@@ -24,7 +24,7 @@ defmodule Openmaize.ConfirmEmailTest do
   end
 
   def user_confirmed do
-    user = TestRepo.get_by(User, email: "fred+1@mail.com")
+    user = TestRepo.get_by(TestUser, email: "fred+1@mail.com")
     user.confirmed_at
   end